adding lombok

+++ src/main/java/com/hariti/asmaa/FranceTour/services/CompetitionService.java
@@ -1,7 +1,7 @@
-package com.hariti.asmaa.FranceTour.Services;
+package com.hariti.asmaa.FranceTour.services;
 
-import com.hariti.asmaa.FranceTour.Entities.Competition;
-import com.hariti.asmaa.FranceTour.Repositories.CompetitionRepository;
+import com.hariti.asmaa.FranceTour.entities.Competition;
+import com.hariti.asmaa.FranceTour.repositories.CompetitionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -37,10 +37,17 @@ public class CompetitionService {
     }
 
     public Competition updateCompetition(Competition competition) {
+        if(competition ==null) { throw new NullPointerException("Competition is null"); }
+        if(!competitionRepository.existsById(competition.getId())) {
+            throw new NullPointerException("Competition not found");
+        }
         return competitionRepository.save(competition);
     }
 
     public void deleteCompetition(long id) {
+        if(!competitionRepository.existsById(id)) {
+            throw new NullPointerException("Competition not found");
+        }
         competitionRepository.deleteById(id);
     }
 }
